toutes les fonctions sont corriges

+++ ft_strlen.c
@@ -6,7 +6,7 @@
 /*   By: rosanche <marvin@42.fr>                    +#+  +:+       +#+        */
 /*                                                +#+#+#+#+#+   +#+           */
 /*   Created: 2017/11/19 19:15:36 by rosanche          #+#    #+#             */
-/*   Updated: 2017/12/07 20:06:30 by rosanche         ###   ########.fr       */
+/*   Updated: 2017/12/08 15:02:33 by rosanche         ###   ########.fr       */
 /*                                                                            */
 /* ************************************************************************** */
 
@@ -14,7 +14,7 @@
 
 int	ft_strlen(const char *str)
 {
-	int i;
+	size_t i;
 
 	i = 0;
 	while (str[i])
